+++ amqp-open-20191212/alibabacloud_amqp_open20191212/models/_list_instances_response_body.py
@@ -112,6 +112,7 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         instance_name: str = None,
         instance_type: str = None,
         kms_key_id: str = None,
+        listener_mode: str = None,
         max_eip_tps: int = None,
         max_queue: int = None,
         max_tps: int = None,
@@ -122,10 +123,14 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         provisioned_capacity: int = None,
         public_endpoint: str = None,
         resource_group_id: str = None,
+        security_group_id: str = None,
+        serverless_switch: bool = None,
         status: str = None,
         storage_size: int = None,
         support_eip: bool = None,
         tags: List[main_models.ListInstancesResponseBodyDataInstancesTags] = None,
+        vpc_id: str = None,
+        vswitch_ids: List[str] = None,
     ):
         # Indicates whether the instance is automatically renewed.
         self.auto_renew_instance = auto_renew_instance
@@ -148,6 +153,7 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         self.instance_type = instance_type
         # The ID of the Key Management Service (KMS) key used for the data disk.
         self.kms_key_id = kms_key_id
+        self.listener_mode = listener_mode
         # The maximum number of Internet-based transactions per second (TPS) for the instance.
         self.max_eip_tps = max_eip_tps
         # The maximum number of queues on the instance.
@@ -170,6 +176,8 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         self.public_endpoint = public_endpoint
         # The ID of the resource group to which the instance belongs.
         self.resource_group_id = resource_group_id
+        self.security_group_id = security_group_id
+        self.serverless_switch = serverless_switch
         # The instance status. Valid values:
         # 
         # *   DEPLOYING: The instance is being deployed.
@@ -185,6 +193,8 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         self.support_eip = support_eip
         # The tags that are added to the instance.
         self.tags = tags
+        self.vpc_id = vpc_id
+        self.vswitch_ids = vswitch_ids
 
     def validate(self):
         if self.tags:
@@ -224,6 +234,9 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         if self.kms_key_id is not None:
             result['KmsKeyId'] = self.kms_key_id
 
+        if self.listener_mode is not None:
+            result['ListenerMode'] = self.listener_mode
+
         if self.max_eip_tps is not None:
             result['MaxEipTps'] = self.max_eip_tps
 
@@ -254,6 +267,12 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         if self.resource_group_id is not None:
             result['ResourceGroupId'] = self.resource_group_id
 
+        if self.security_group_id is not None:
+            result['SecurityGroupId'] = self.security_group_id
+
+        if self.serverless_switch is not None:
+            result['ServerlessSwitch'] = self.serverless_switch
+
         if self.status is not None:
             result['Status'] = self.status
 
@@ -267,6 +286,12 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         if self.tags is not None:
             for k1 in self.tags:
                 result['Tags'].append(k1.to_map() if k1 else None)
+
+        if self.vpc_id is not None:
+            result['VpcId'] = self.vpc_id
+
+        if self.vswitch_ids is not None:
+            result['VswitchIds'] = self.vswitch_ids
 
         return result
 
@@ -299,6 +324,9 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         if m.get('KmsKeyId') is not None:
             self.kms_key_id = m.get('KmsKeyId')
 
+        if m.get('ListenerMode') is not None:
+            self.listener_mode = m.get('ListenerMode')
+
         if m.get('MaxEipTps') is not None:
             self.max_eip_tps = m.get('MaxEipTps')
 
@@ -329,6 +357,12 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
         if m.get('ResourceGroupId') is not None:
             self.resource_group_id = m.get('ResourceGroupId')
 
+        if m.get('SecurityGroupId') is not None:
+            self.security_group_id = m.get('SecurityGroupId')
+
+        if m.get('ServerlessSwitch') is not None:
+            self.serverless_switch = m.get('ServerlessSwitch')
+
         if m.get('Status') is not None:
             self.status = m.get('Status')
 
@@ -343,6 +377,12 @@ class ListInstancesResponseBodyDataInstances(DaraModel):
             for k1 in m.get('Tags'):
                 temp_model = main_models.ListInstancesResponseBodyDataInstancesTags()
                 self.tags.append(temp_model.from_map(k1))
+
+        if m.get('VpcId') is not None:
+            self.vpc_id = m.get('VpcId')
+
+        if m.get('VswitchIds') is not None:
+            self.vswitch_ids = m.get('VswitchIds')
 
         return self